improve stats api tests
