+++ txbits/test/EngineModelSpec.scala
@@ -293,75 +293,5 @@ class EngineModelSpec extends Specification with Mockito {
         }
       }
     }
-
-    "update stats after a bid matches" in new WithCleanTestDbApplication {
-      val asker = globals.userModel.create("test@test.test", "", false).get
-      val bidder = globals.userModel.create("test2@test.test", "", false).get
-
-      globals.userModel.addFakeMoney(asker, "LTC", 1)
-      globals.userModel.addFakeMoney(bidder, "USD", 1)
-
-      globals.engineModel.askBid(Some(asker), None, "LTC", "USD", 1, 1, false)
-      globals.engineModel.askBid(Some(bidder), None, "LTC", "USD", 1, 1, true)
-
-      val stats_res = controllers.StatsAPI.APIv1.chartFromDB("LTC", "USD")
-      stats_res.head(1).value should be equalTo 1
-      stats_res.head(2).value should be equalTo 1
-      stats_res.head(3).value should be equalTo 1
-      stats_res.head(4).value should be equalTo 1
-      stats_res.head(5).value should be equalTo 1
-    }
-
-    "update stats after an ask matches" in new WithCleanTestDbApplication {
-      val asker = globals.userModel.create("test@test.test", "", false).get
-      val bidder = globals.userModel.create("test2@test.test", "", false).get
-
-      globals.userModel.addFakeMoney(asker, "LTC", 1)
-      globals.userModel.addFakeMoney(bidder, "USD", 1)
-
-      globals.engineModel.askBid(Some(bidder), None, "LTC", "USD", 1, 1, true)
-      globals.engineModel.askBid(Some(asker), None, "LTC", "USD", 1, 1, false)
-
-      val stats_res = controllers.StatsAPI.APIv1.chartFromDB("LTC", "USD")
-      stats_res.head(1).value should be equalTo 1
-      stats_res.head(2).value should be equalTo 1
-      stats_res.head(3).value should be equalTo 1
-      stats_res.head(4).value should be equalTo 1
-      stats_res.head(5).value should be equalTo 1
-    }
-
-    "not update stats for no matches" in new WithCleanTestDbApplication {
-      val asker = globals.userModel.create("test@test.test", "", false).get
-      val bidder = globals.userModel.create("test2@test.test", "", false).get
-
-      globals.userModel.addFakeMoney(asker, "LTC", 1)
-      globals.userModel.addFakeMoney(bidder, "USD", 1)
-
-      globals.engineModel.askBid(Some(asker), None, "LTC", "USD", 1, 2, false)
-      globals.engineModel.askBid(Some(bidder), None, "LTC", "USD", 1, 1, true)
-
-      val stats_res = controllers.StatsAPI.APIv1.chartFromDB("LTC", "USD")
-      stats_res should beEmpty
-    }
-
-    "update stats correctly after 3 matches" in new WithCleanTestDbApplication {
-      val asker = globals.userModel.create("test@test.test", "", false).get
-      val bidder = globals.userModel.create("test2@test.test", "", false).get
-
-      globals.userModel.addFakeMoney(asker, "LTC", 10)
-      globals.userModel.addFakeMoney(bidder, "USD", 10)
-
-      globals.engineModel.askBid(Some(asker), None, "LTC", "USD", 1, 1, false)
-      globals.engineModel.askBid(Some(bidder), None, "LTC", "USD", 2, 2, true)
-      globals.engineModel.askBid(Some(asker), None, "LTC", "USD", 2, 1.5, false)
-      globals.engineModel.askBid(Some(bidder), None, "LTC", "USD", 1, 1.5, true)
-
-      val stats_res = controllers.StatsAPI.APIv1.chartFromDB("LTC", "USD")
-      stats_res.head(1).value should be equalTo 1
-      stats_res.head(2).value should be equalTo 2
-      stats_res.head(3).value should be equalTo 1
-      stats_res.head(4).value should be equalTo 1.5
-      stats_res.head(5).value should be equalTo 3
-    }
   }
 }
